Work on different US
- US41-Create new Play Sudoku Class instead menu
- US38-Implement last good square in console
- US37-Implement Reset method in console
- US36-Implement Hint method in console
- US33-Undo function for Play sudoku

+++ src/Main/play_sudoku.py
@@ -94,7 +94,7 @@ class PlaySudoku:
         
     def __calculate_time(self,parsed_time):
         minutes,seconds = parsed_time.split(":")
-        return ((int(minutes)*60) + int(seconds))
+        return ((int(minutes)*60) + float(seconds))
         
     def get_time(self):
         '''
@@ -116,7 +116,7 @@ class PlaySudoku:
         '''
         self.initial_time = new_time
         minutes,seconds = new_time.split(":")
-        self.time_elapsed = (int(minutes)*60) + int(seconds)
+        self.time_elapsed = (int(minutes)*60) + float(seconds)
     
     def __save(self):
         '''
